Add the payment logos
This isn't quite ready to merge yet – I need to add licence/copyright
information for each of the payment provider icons and possibly link to
more legal info.

The structure of this repo is mostly copied from flag images. The
CircleCI config is new, and will be applied to all of the image sets
once I verify that it's working here.